pyproject.toml
[project]
name = "stock-data-pipeline"
version = "0.1.0"
description = ""
authors = [
    {name = "ErEr",email = "jeffliao92316046@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "apache-airflow (>=2.10.4,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "sqlalchemy (>=1.4.36,<2.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false